% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BF2posterior}
\alias{BF2posterior}
\title{convert BF to posterior estimates}
\usage{
BF2posterior(
  BF,
  Genes = names(BF),
  prior = predict_uniform_prior(BF, Genes),
  log_BF = FALSE
)
}
\arguments{
\item{BF}{bayes factor (or log bayes factor}

\item{Genes}{character vector for gene names}

\item{prior}{prior (or log prior)}

\item{log_BF}{whether to compute in log space (and return in log space)}
}
\value{

}
\description{
convert BF to posterior estimates
}
