% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_Func.R
\name{fgem}
\alias{fgem}
\title{Run FGEM}
\usage{
fgem(
  X,
  BF,
  prior = 0.02,
  tmu = prior_mean(BF, prior),
  null_beta = fgem_null(BF, tmu),
  verbose = FALSE
)
}
\arguments{
\item{X}{matrix of gene-level annotations}

\item{BF}{vector of gene-level bayes factors}

\item{prior}{prior probability a gene is causal (before annotations)}

\item{tmu}{precomputed gene-level prior}

\item{null_beta}{model to test against for likelihood ratio test}

\item{verbose}{whether to use verbose output}
}
\value{
tibble with fgem results
}
\description{
Run FGEM
}
