% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgem-package.R
\docType{package}
\name{fgem-package}
\alias{fgem}
\alias{fgem-package}
\title{fgem: R package for Functional Genomics using Expectation Maximization (FGEM)}
\description{
Given a list of gene-level bayes factors and a series of annotations about those genes, FGEM simultaneously calculates a posterior probability of genes being causal, and estimates the effect sizeof the annotations.
}
\keyword{internal}
