% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_Func.R
\name{fgem}
\alias{fgem}
\title{Run FGEM}
\usage{
fgem(
  x,
  BF,
  alpha = 1,
  lambda = c(2, 1, 5 * 10^(-(seq(1, 6, length.out = 75))), 0),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{BF}{vector of gene-level bayes factors}

\item{verbose}{whether to use verbose output}

\item{X}{matrix of gene-level annotations}
}
\value{
tibble with fgem results
}
\description{
Run FGEM
}
