% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fgem_grad}
\alias{fgem_grad}
\alias{fgem_grad.matrix}
\alias{fgem_grad.dgCMatrix}
\alias{fgem_grad.default}
\title{Calculate fgem gradient}
\usage{
fgem_grad(par, X, BF, l2 = 0, l1 = 0, log_BF = FALSE, ...)

\method{fgem_grad}{matrix}(par, X, BF, l2 = 0, l1 = 0, log_BF = FALSE, ...)

\method{fgem_grad}{dgCMatrix}(par, X, BF, l2 = 0, l1 = 0, log_BF = FALSE, ...)

\method{fgem_grad}{default}(par, X, BF, l2 = 0, l1 = 0, log_BF = FALSE, ...)
}
\arguments{
\item{par}{Beta parameter}

\item{X}{matrix (can be Matrix from Matrix package)}

\item{BF}{bayes factor (or log Bayes factor if log_BF=TRUE)}

\item{l2}{L2 regularization}

\item{log_BF}{boolean indicating whether BF should be treated as log_BF}
}
\value{
vector with gradients
}
\description{
Calculate fgem gradient
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: gradient for dense matrix

\item \code{dgCMatrix}: gradient for sparse matrix

\item \code{default}: gradient for generic
}}

