% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fgem_grad_stan}
\alias{fgem_grad_stan}
\title{Gradient for FGEM likelihood
This is an attempt to use stan's AD features to optimize the FGEM likelihood}
\usage{
fgem_grad_stan(par, X, BF, l2 = 0, l1 = 0, log_BF = FALSE)
}
\description{
Gradient for FGEM likelihood
This is an attempt to use stan's AD features to optimize the FGEM likelihood
}
