% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fgem_hess}
\alias{fgem_hess}
\title{Calculate fgem gradient}
\usage{
fgem_hess(par, X, BF, l2 = 0, l1 = 0, log_BF = FALSE, ...)
}
\arguments{
\item{par}{Beta parameter}

\item{X}{matrix (can be Matrix from Matrix package)}

\item{BF}{bayes factor (or log Bayes factor if log_BF=TRUE)}

\item{l2}{L2 regularization}

\item{log_BF}{boolean indicating whether BF should be treated as log_BF}
}
\value{
vector with gradients
}
\description{
Calculate fgem gradient
}
