% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_Func.R
\name{fgem_lik}
\alias{fgem_lik}
\title{Maximize FGEM marginalized likelihood}
\usage{
fgem_lik(
  x,
  BF,
  prior = 0.02,
  tmu = prior_mean(BF, prior),
  add_intercept = TRUE,
  Beta0 = guess_beta0(x, BF, prior, tmu, add_intercept),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{gene-level feature set}

\item{BF}{gene-level bayes factors}

\item{prior}{(scalar?) prior probability z=1}

\item{tmu}{(gene-level) prior probability z=1 (takes precedence over `prior` if both are specified)}

\item{Beta0}{initial guess for enrichment}

\item{verbose}{logical scalar.  Print out lots of debug info?}
}
\value{
dataframe with nested dataframe with multivariate effect-size estimates, and FGEM likelihood
}
\description{
Maximize FGEM marginalized likelihood
}
