% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_Func.R
\name{FGEM_marginal}
\alias{FGEM_marginal}
\title{Run FGEM with a dataframe as input
This version of FGEM takes a single dataframe with
both gene-level bayes factors and gene-level features
and returns a dataframe with effect size estimates
for every feature (run as univariate)}
\usage{
FGEM_marginal(X, BF, prior_mean = 0.02, verbose = FALSE, parallel = FALSE, ...)
}
\arguments{
\item{prior_mean}{scalar between 0 and 1 specifying the starting
prior probability (this is for initializing the EM algorithm).
If unspecified, it defaults to 0.02}

\item{verbose}{Whether to print debug output}

\item{feat_df}{dataframe with data. `feat_df`
must have a column called `BF` specifying bayes factors.
  Any additional columns (besides an optional gene-name
column that must be titled `Gene`) are treated as annotations.}
}
\description{
Run FGEM with a dataframe as input
This version of FGEM takes a single dataframe with
both gene-level bayes factors and gene-level features
and returns a dataframe with effect size estimates
for every feature (run as univariate)
}
