% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_Func.R
\name{gen_posterior}
\alias{gen_posterior}
\title{Generate posterior from fitted FGEM_df result and annotation dataframe}
\usage{
gen_posterior(feat_df, fgem_result_df)
}
\arguments{
\item{feat_df}{annotation dataframe (as described inf `FGEM_df`)}

\item{fgem_result_df}{result of running FGEM_df}
}
\description{
Generate posterior from fitted FGEM_df result and annotation dataframe
}
