% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_Func.R
\name{gen_prior}
\alias{gen_prior}
\title{Generate prior from fitted FGEM_df result and annotation dataframe}
\usage{
gen_prior(feat_df, fgem_result_df)
}
\arguments{
\item{feat_df}{annotation dataframe (as described inf `FGEM_df`)}

\item{fgem_result_df}{result of running FGEM_df}
}
\description{
Generate prior from fitted FGEM_df result and annotation dataframe
}
