% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log1mexp}
\alias{log1mexp}
\title{Compute  f(a) = log(1 - exp(-a))  stably}
\usage{
log1mexp(x)
}
\arguments{
\item{a}{numeric vector of positive values}

\item{cutoff}{log(2) is optimal, see  Maechler (201x) .....}
}
\value{
f(a) == log(1 - exp(-a)) == log1p(-exp(-a)) == log(-expm1(-a))
}
\description{
Compute  f(a) = log(1 - exp(-a))  stably
}
\author{
Martin Maechler, May 2002 .. Aug. 2011
}
