% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_pbernoulli}
\alias{log_pbernoulli}
\title{calculate log loss for log-valued prediction}
\usage{
log_pbernoulli(lp, x)
}
\arguments{
\item{lp}{(natural) log-scale probablilty values}

\item{x}{integer (or logical) of length equal to lp indicating}
}
\value{

}
\description{
calculate log loss for log-valued prediction
}
