% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lseq}
\alias{lseq}
\title{Sequence generation in log-space}
\usage{
lseq(from = 1, to = 1e+05, length.out = 6, base = 10)
}
\arguments{
\item{from}{starting point}

\item{to}{maximal end value}

\item{length.out}{desired length of the sequence}

\item{base}{log-base (default is 10)}
}
\value{

}
\description{
Sequence generation in log-space
}
