% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{marginal_fgem_fit_bfgs}
\alias{marginal_fgem_fit_bfgs}
\title{fit every column of X with the BFGS algorithm}
\usage{
marginal_fgem_fit_bfgs(
  X,
  BF,
  prec = 0,
  epsilon = 1e-06,
  max_iter = 100L,
  progress = TRUE
)
}
\description{
fit fgem using c++ BFGS algorithm
}
