% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_Func.R
\name{partial_resolve}
\alias{partial_resolve}
\title{Helper to get resolved futures}
\usage{
partial_resolve(x)
}
\arguments{
\item{x}{an object}
}
\value{
if x inherits Future, it returns
the value of x if it is resolved, or itself
if is not.  If x is not a Future, it returns
itself
}
\description{
Helper to get resolved futures
}
