% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_Func.R
\name{resolved_or_null}
\alias{resolved_or_null}
\title{Helper to get resolved futures}
\usage{
resolved_or_null(x)
}
\arguments{
\item{x}{an object}
}
\value{
if x inherits Future, it returns
the value of x if it is resolved, or NULL
if is not.  If x is not a Future, it returns
x
}
\description{
Helper to get resolved futures
}
