% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_Func.R
\name{retrieve_resolved}
\alias{retrieve_resolved}
\title{retrieve only resolved results from list of futures}
\usage{
retrieve_resolved(x, atleast = 1)
}
\arguments{
\item{x}{a list of futures}

\item{atleast}{the min number of resolved futures to return}
}
\value{
a list of values of length <= length(x)
}
\description{
retrieve only resolved results from list of futures
}
