% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sp_fgem_hess_stan}
\alias{sp_fgem_hess_stan}
\title{evd_dnorm_hess_stan}
\usage{
sp_fgem_hess_stan(par, X, BF, l2 = 0, l1 = 0, neg = FALSE, log_BF = FALSE)
}
\description{
This is an attempt to use stan's AD features to calculate a hessian
for the RSSp likelihood
}
