% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sp_fgem_lik_stan}
\alias{sp_fgem_lik_stan}
\title{This is an attempt to use stan's AD features to calculate a gradient
for the FGEM likelihood}
\usage{
sp_fgem_lik_stan(par, X, BF, prec = 0, neg = FALSE, log_BF = FALSE)
}
\description{
This is an attempt to use stan's AD features to calculate a gradient
for the FGEM likelihood
}
