% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_Func.R
\name{trip2sparseMatrix}
\alias{trip2sparseMatrix}
\title{Convert a table-style sparse matrix representation to a sparse matrix}
\usage{
trip2sparseMatrix(
  rowname_vals,
  colname_vals,
  values = rep(1, length(colname_vals)),
  total_rownames = unique(rowname_vals),
  total_colnames = unique(colname_vals),
  add_intercept = TRUE,
  csr = FALSE
)
}
\arguments{
\item{rowname_vals}{vector with the rownames of the values}

\item{colname_vals}{vector with the colnames of the values}

\item{values}{the values themselves (default is 1.0)}

\item{total_rownames}{the universe of rownames, defaults to all the unique rowname_vals}

\item{total_colnames}{the universe of colnames, defaults to all the unique colname_vals}

\item{add_intercept}{whether to add an intercept term ("Intercept" will be added to total_colnames)}
}
\value{
a sparse matrix of dim(length(total_rownames),length(total_colnames)) with (at most) length(values)
non-zero values
}
\description{
Convert a table-style sparse matrix representation to a sparse matrix
}
