% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_Func.R
\name{FGEM}
\alias{FGEM}
\title{Run FGEM with a matrix as input


This version of FGEM takes a single dataframe with
both gene-level bayes factors and gene-level features
and returns a dataframe with effect size estimates
for every feature.}
\usage{
FGEM(Beta0, feat_mat, BF, verbose = FALSE, null_beta = NULL, ...)
}
\arguments{
\item{Beta0}{initial Beta0}

\item{feat_mat}{matrix of features}

\item{BF}{vector of bayes factors}

\item{verbose}{whether to give print intermediate results to the user}

\item{null_beta}{Instead of refitting the null model, you can pass the
precomputed beta for the intercept-only model}

\item{...}{currently unused}

\item{null_features}{which features are "null" (deprecated)}
}
\value{
results of model fit as a tibble
}
\description{
Run FGEM with a matrix as input


This version of FGEM takes a single dataframe with
both gene-level bayes factors and gene-level features
and returns a dataframe with effect size estimates
for every feature.
}
