% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_Func.R
\name{FGEM}
\alias{FGEM}
\title{Run FGEM with a matrix as input
This version of FGEM takes a single dataframe with 
both gene-level bayes factors and gene-level features 
and returns a dataframe with effect size estimates 
for every feature.}
\usage{
FGEM(Beta0, feat_mat, BF, null_features = "Intercept")
}
\arguments{
\item{Beta0}{vector specifying starting guesses for effect sizes of annotations. 
In almost all cases the default (NULL) should be kept}

\item{null_features}{this is a character vector specifying which features should be included in the null model.
The default is to only include an intercept term.}

\item{feat_df}{dataframe with data.  `feat_df` must have a column called `BF` specifying bayes factors.
Any additional columns are treated as annotations.}

\item{prior_mean}{scalar between 0 and 1 specifying the starting 
prior probability (this is for initializing the EM algorithm). If unspecified, it defaults to 0.2}
}
\description{
Run FGEM with a matrix as input
This version of FGEM takes a single dataframe with 
both gene-level bayes factors and gene-level features 
and returns a dataframe with effect size estimates 
for every feature.
}
