% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fgem_lik}
\alias{fgem_lik}
\alias{fgem_lik.matrix}
\alias{fgem_lik.dgCMatrix}
\alias{fgem_lik.default}
\title{Calculate fgem likelihood}
\usage{
fgem_lik(par, X, BF, l2 = 0, l1 = 0, log_BF = FALSE, ...)

\method{fgem_lik}{matrix}(par, X, BF, l2 = 0, l1 = 0, log_BF = FALSE, ...)

\method{fgem_lik}{dgCMatrix}(par, X, BF, l2 = 0, l1 = 0, log_BF = FALSE, ...)

\method{fgem_lik}{default}(par, X, BF, l2 = 0, l1 = 0, log_BF = FALSE, ...)
}
\arguments{
\item{par}{Beta parameter}

\item{X}{matrix (can be Matrix from Matrix package)}

\item{BF}{bayes factor (or log Bayes factor if log_BF=TRUE)}

\item{l2}{L2 regularization}

\item{log_BF}{boolean indicating whether BF should be treated as log_BF}
}
\value{
vector with gradients
}
\description{
Calculate fgem likelihood
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: likelihood for dense data matrix

\item \code{dgCMatrix}: likelihood for sparse data matrix

\item \code{default}: likelihood
}}

