% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_Func.R
\name{fgem_marginal}
\alias{fgem_marginal}
\title{Run FGEM with a dataframe as input
This version of FGEM takes a single dataframe with
both gene-level bayes factors and gene-level features
and returns a dataframe with effect size estimates
for every feature (run as univariate)}
\usage{
fgem_marginal(X, BF, prior_mean = 0.02, epsilon = 1e-06, max_iter = 150, ...)
}
\arguments{
\item{X}{feature matrix (must have column names)}

\item{BF}{vector of bayes factors}

\item{verbose}{Whether to print debug output}
}
\description{
Run FGEM with a dataframe as input
This version of FGEM takes a single dataframe with
both gene-level bayes factors and gene-level features
and returns a dataframe with effect size estimates
for every feature (run as univariate)
}
