% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict_fgem}
\alias{predict_fgem}
\title{Predict posterior from fgem fit}
\usage{
predict_fgem(fit, x, log = FALSE)
}
\arguments{
\item{fit}{result of call to `fgem`}

\item{x}{annotation matrix (colnames should correspond to `feature_name` column of fit$Beta[[1]])}

\item{log}{return log posterior}

\item{BF}{Bayes Factor}

\item{lambda}{value of lambda to use in prediction}
}
\value{
vector of length length(BF) with posterior probabilities
}
\description{
Predict posterior from fgem fit
}
